+++ cSTIR/cstir_p.h
