+++ cSTIR/cstir_tw.cpp
